change from having personal teams to creating team at register
Jetstream's default was to create a personal team for the user. This
could end up creating a lot of default teams in our SaaS application that
are not being used.

Instead, we have the user create their team on registration. Middleware
and a trait were also added to ensure that a user does have a team before
continuing so that we do not have issues with a team not being set.

When a user hits a route with the teams middleware, we check to see
if they have a current_team_id set, if so continue. If not, we check
to see if they belong to a team and set that as their current team.
If the user does not belong to a team, we will send them to the create
team page.

Signed-off-by: Daniel Berry <daniel@jstreetdigital.com>